Refactored graphql schemas
