+++ docs/_docs/dev-guide/tavern.md
@@ -108,31 +108,29 @@ If you'd like to explore the Graph API and try out some queries, head to the `/g
 ## Creating a New Model
 1. Initialize the schema `cd tavern && go run entgo.io/ent/cmd/ent init <NAME>`
 2. Update the generated file in `tavern/ent/schema/<NAME>.go`
-3. Ensure you include a `func (<NAME>) Annotations() []schema.Annotation` method which returns a `entgql.QueryField()` annotation to tell entgo to generate a GraphQL root query for this model (if you'd like it to be queryable from the root query).
+3. Ensure you include a `func (<NAME>) Annotations() []schema.Annotation` method which returns a `entgql.QueryField()` annotation to tell entgo to generate a GraphQL root query for this model (if you'd like it to be queryable from the root query)
 4. Update `tavern/graphql/gqlgen.yml` to include the ent types in the `autobind:` section (e.g.`- github.com/kcarretto/realm/tavern/ent/<NAME>`)
 5. **Optionally** update the `models:` section of `tavern/graphql/gqlgen.yml` to bind any GraphQL enum types to their respective `entgo` generated types (e.g. `github.com/kcarretto/realm/tavern/ent/<NAME>.<ENUM_FIELD>`)
 6. Run `go generate ./tavern/...` from the project root
-    * **Note:** There is currently a bug with entgo & gqlgen where you may sometimes get the error `failed to load schema: schema/schema.graphql:15: Undefined type Time.`, this is usually fixed by running `go generate ./tavern/...` again
-7. You will notice auto-generated go "resolver" files have been updated with new methods to query your model (e.g. `func (r *queryResolver) <NAME>s ...`)
+7. If you added an annotation for a root query field (see above), you will notice auto-generated the `query.resolvers.go` file has been updated with new methods to query your model (e.g. `func (r *queryResolver) <NAME>s ...`)
     * This must be implemented (e.g. `return r.client.<NAME>.Query().All(ctx)` where <NAME> is the name of your model)
 
 ## Adding Mutations
-1. Create a new GraphQL schema file for your model if it does not yet exist (e.g. `touch ./tavern/graphql/schema/<NAME>.graphql` from the project root)
-    * If it did not already exist, update `tavern/graphql/gqlgen.yml` to include this schema under the `schema:` section.
-2. Add any mutations to `tavern/graphql/schema/<NAME>.graphql`
+1. Update the `mutation.graphql` schema file to include your new mutation and please include it in the section for the model it's mutating if applicable (e.g. createUser should be defined near all the related User mutations)
     * **Note:** Input types such as `Create<NAME>Input` or `Update<NAME>Input` will already be generated if you [added the approproate annotations to your ent schema](https://entgo.io/docs/tutorial-todo-gql#install-and-configure-entgql)
-3. Run `go generate ./...`
-4. Implement generated the generated mutation resolver method in `tavern/graphql/<NAME>.resolvers.go`
-    * Depending on the mutation you're trying to implement, a one liner such as `return r.client.User.Create().SetInput(input).Save(ctx)` might be sufficient
-5. Please write a unit test for your new mutation <3
+2. Run `go generate ./...`
+3. Implement generated the generated mutation resolver method in `tavern/graphql/mutation.resolvers.go`
+    * Depending on the mutation you're trying to implement, a one liner such as `return r.client.<NAME>.Create().SetInput(input).Save(ctx)` might be sufficient
+5. Please write a unit test for your new mutation in `mutation_test.go` <3
 
 ## Code Generation Reference
 * After making a change, remember to run `go generate ./...` from the project root.
 * `tavern/ent/schema` is a directory which defines our graph using database models (ents) and the relations between them
 * `tavern/generate.go` is responsible for generating ents defined by the ent schema as well as updating the GraphQL schema and generating related code
 * `tavern/ent/entc.go` is responsible for generating code for the entgo <-> 99designs/gqlgen GraphQL integration
-* `tavern/graphql/schema/schema.graphql` is a GraphQL schema automatically generated by ent, providing useful types derived from our ent schemas
-* `tavern/graphql/schema/<NAME>.graphql` defines mutations for the given entity
+* `tavern/graphql/schema/mutation.graphql` defines all mutations supported by our API
+* `tavern/graphql/schema/query.graphql` is a GraphQL schema automatically generated by ent, providing useful types derived from our ent schemas as well as root-level queries defined by entgo annotations
+* `tavern/graphql/schema/scalars.graphql` defines scalar GraphQL types that can be used to help with Go bindings (See [gqlgen docs](https://gqlgen.com/reference/scalars/) for more info)
 
 ## Resources
 * [Relay Documentation](https://relay.dev/graphql/connections.htm)
