+++ tavern/ent/schema/job.go
@@ -43,5 +43,8 @@ func (Job) Edges() []ent.Edge {
 func (Job) Annotations() []schema.Annotation {
 	return []schema.Annotation{
 		entgql.QueryField(),
+		entgql.Mutations(
+			entgql.MutationCreate(),
+		),
 	}
 }
