+++ tavern/graphql/schema/mutation.graphql
@@ -1,5 +1,8 @@
 type Mutation {
-    # The input and the output are types generated by Ent.
+
+    ### 
+    # User
+    ###  
     createUser(input: CreateUserInput!): User
     updateUser(id: ID!, input: UpdateUserInput!): User
 }
