+++ tavern/graphql/schema/query.graphql
@@ -1,6 +1,17 @@
 directive @goField(forceResolver: Boolean, name: String) on FIELD_DEFINITION | INPUT_FIELD_DEFINITION
 directive @goModel(model: String, models: [String!]) on OBJECT | INPUT_OBJECT | SCALAR | ENUM | INTERFACE | UNION
 """
+CreateJobInput is used for create Job object.
+Input was generated by ent.
+"""
+input CreateJobInput {
+  """Name of the job"""
+  name: String!
+  createdbyID: ID!
+  tomeID: ID!
+  taskIDs: [ID!]
+}
+"""
 CreateUserInput is used for create User object.
 Input was generated by ent.
 """
