+++ tavern/internal/www/schema.graphql
@@ -1,5 +1,23 @@
-scalar Timedirective @goField(forceResolver: Boolean, name: String) on FIELD_DEFINITION | INPUT_FIELD_DEFINITION
+type Mutation {
+
+    ### 
+    # User
+    ###  
+    createUser(input: CreateUserInput!): User
+    updateUser(id: ID!, input: UpdateUserInput!): User
+}directive @goField(forceResolver: Boolean, name: String) on FIELD_DEFINITION | INPUT_FIELD_DEFINITION
 directive @goModel(model: String, models: [String!]) on OBJECT | INPUT_OBJECT | SCALAR | ENUM | INTERFACE | UNION
+"""
+CreateJobInput is used for create Job object.
+Input was generated by ent.
+"""
+input CreateJobInput {
+  """Name of the job"""
+  name: String!
+  createdbyID: ID!
+  tomeID: ID!
+  taskIDs: [ID!]
+}
 """
 CreateUserInput is used for create User object.
 Input was generated by ent.
@@ -605,8 +623,4 @@ input UserWhereInput {
   isadmin: Boolean
   isadminNEQ: Boolean
 }
-type Mutation {
-    # The input and the output are types generated by Ent.
-    createUser(input: CreateUserInput!): User
-    updateUser(id: ID!, input: UpdateUserInput!): User
-}
+scalar Time
